mcp-server/requirements.txt
mcp[cli]>=1.2.0
httpx>=0.27.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.30.0
